+++ test/dummy/app/models/user.rb
@@ -2,6 +2,10 @@ class User < ActiveRecord::Base
   include ActiveModel::ForbiddenAttributesProtection
   include Authorizable::User  
 
+  def admin?
+    role == 'admin'
+  end
+
   def access_roles
     [].tap do |allowed|
       allowed << :admin if admin?